requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-filter>=23.0
channels>=4.0
channels-redis>=4.1
celery>=5.3
redis>=5.0
python-dotenv>=1.0
reportlab>=4.0
